+++ src/main/java/example/service/repository/PriceDefinitionRepository.java
@@ -4,6 +4,6 @@ import example.service.model.PriceDefinition;
 import java.util.List;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PriceDefinitionRepository extends CrudRepository<PriceDefinition, Long> {
+public interface PriceDefinitionRepository extends CrudRepository<PriceDefinition, Integer> {
     List<PriceDefinition> findByBillingPlan_Id(Integer billingPlanId);
 }
